pam_ssh_nullok fix
Date: Tue, 15 Nov 2011 22:12:46 +0100
From: Dag-Erling Smorgrav <des@des.no>
To: Guy Helmer <guy.helmer@palisadesystems.com>
Cc: freebsd-security@freebsd.org
Subject: Re: Possible pam_ssh bug?

It turns out that this goes all the way to OpenSSL, which ignores the
passphrase if the key is not encrypted.  The only solution I can think
of - more of a workaround, really - is to first try to load the key with
an empty passphrase, and skip the key if that worked.  See the attached
(untested) patch.

A more advanced patch would load all keys but require at least one of
them to have a passphrase.